+++ com.zerogamestudio.zeroengine.pathfinding2d/Runtime/AssemblyInfo.cs
@@ -1,0 +1,3 @@
+using System.Runtime.CompilerServices;
+
+[assembly: InternalsVisibleTo("ZeroEngine.Pathfinding2D.Tests.Editor")]